openWeather plugin for openLuup
This plug-in is intended to run under the "openLuup" emulation of a Vera
system. It should work on a "real" Vera, but has not been tested in that
environment. It is intended to capture and monitor select weather data
provided by Weather Underground www.weatherunderground.com under their
general terms and conditions available on their website. It requires an
API developer key that must be obtained from their website.